requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
absl-py==0.11.0           # via tensorboard, tensorflow
argcomplete==1.10.0       # via textract
astunparse==1.6.3         # via tensorflow
backcall==0.2.0           # via ipython
beautifulsoup4==4.8.0     # via bs4, textract
blis==0.4.1               # via spacy, thinc
bs4==0.0.1                # via -r requirements.in
cachetools==4.1.1         # via google-auth
catalogue==1.0.0          # via spacy, thinc
certifi==2020.11.8        # via requests
chardet==3.0.4            # via -r requirements.in, requests, textract
click==7.1.2              # via nltk, sacremoses
contractions==0.0.25      # via -r requirements.in
cycler==0.10.0            # via matplotlib
cymem==2.0.4              # via preshed, spacy, thinc
dataclasses==0.8          # via torch
decorator==4.4.2          # via ipython, traitlets
defusedxml==0.6.0         # via odfpy
diff-match-patch==20200713  # via django-import-export
dj-database-url==0.5.0    # via -r requirements.in
django-cors-headers==3.5.0  # via -r requirements.in
django-import-export==2.4.0  # via -r requirements.in
django-js-reverse==0.9.1  # via -r requirements.in
django-log-request-id==1.6.0  # via -r requirements.in
django-model-utils==4.0.0  # via -r requirements.in
django-webpack-loader==0.7.0  # via -r requirements.in
django==2.2.17            # via -r requirements.in, django-cors-headers, django-import-export, django-js-reverse, django-log-request-id, django-model-utils, djangorestframework
djangorestframework==3.12.2  # via -r requirements.in
docx2txt==0.8             # via textract
ebooklib==0.17.1          # via textract
et-xmlfile==1.0.1         # via openpyxl
extract-msg==0.23.1       # via textract
future==0.18.2            # via torch
gast==0.3.3               # via tensorflow
google-auth-oauthlib==0.4.2  # via tensorboard
google-auth==1.23.0       # via google-auth-oauthlib, tensorboard
google-pasta==0.2.0       # via tensorflow
grpcio==1.33.2            # via tensorboard, tensorflow
gunicorn==20.0.4          # via -r requirements.in
h5py==2.10.0              # via tensorflow
idna==2.10                # via requests
imapclient==2.1.0         # via extract-msg
importlib-metadata==2.0.0  # via catalogue, markdown
ipython-genutils==0.2.0   # via traitlets
ipython==7.16.1           # via -r requirements.in
jdcal==1.4.1              # via openpyxl
jedi==0.17.2              # via ipython
joblib==0.17.0            # via nltk, sacremoses, scikit-learn
keras-preprocessing==1.1.2  # via tensorflow
kiwisolver==1.3.1         # via matplotlib
langdetect==1.0.8         # via -r requirements.in
laserembeddings==1.1.0    # via -r requirements.in
lxml==4.6.1               # via ebooklib, python-pptx
markdown==3.3.3           # via tensorboard
markuppy==1.14            # via tablib
matplotlib==3.3.3         # via -r requirements.in
murmurhash==1.0.4         # via preshed, spacy, thinc
nltk==3.5                 # via -r requirements.in
numpy==1.18.5             # via -r requirements.in, blis, h5py, keras-preprocessing, laserembeddings, matplotlib, opt-einsum, pandas, scikit-learn, scipy, spacy, tensorboard, tensorflow, tensorflow-hub, thinc, torch, xgboost
oauthlib==3.1.0           # via requests-oauthlib
odfpy==1.4.1              # via tablib
olefile==0.46             # via extract-msg
openpyxl==3.0.5           # via tablib
opt-einsum==3.3.0         # via tensorflow
pandas==1.1.4             # via -r requirements.in
parso==0.7.1              # via jedi
pdfminer.six==20181108    # via textract
pexpect==4.8.0            # via ipython
pickleshare==0.7.5        # via ipython
pillow==8.0.1             # via matplotlib, python-pptx
plac==1.1.3               # via spacy, thinc
plotly==4.12.0            # via -r requirements.in
preshed==3.0.4            # via spacy, thinc
prompt-toolkit==3.0.8     # via ipython
protobuf==3.14.0          # via tensorboard, tensorflow, tensorflow-hub
psutil==5.7.3             # via -r requirements.in
psycopg2-binary==2.8.6    # via -r requirements.in
ptyprocess==0.6.0         # via pexpect
pyahocorasick==1.4.0      # via textsearch
pyasn1-modules==0.2.8     # via google-auth
pyasn1==0.4.8             # via pyasn1-modules, rsa
pycryptodome==3.9.9       # via pdfminer.six
pygments==2.7.2           # via ipython
pymongo==3.11.1           # via -r requirements.in
pyparsing==2.4.7          # via matplotlib
pyphen==0.10.0            # via textstat
python-dateutil==2.8.1    # via matplotlib, pandas
python-decouple==3.3      # via -r requirements.in
python-pptx==0.6.18       # via textract
pytz==2020.4              # via django, pandas, tzlocal
pyyaml==5.3.1             # via tablib
regex==2020.11.13         # via nltk
requests-oauthlib==1.3.0  # via google-auth-oauthlib
requests==2.25.0          # via requests-oauthlib, spacy, tensorboard
retrying==1.3.3           # via plotly
rsa==4.6                  # via google-auth
sacremoses==0.0.35        # via laserembeddings
scikit-learn==0.23.2      # via -r requirements.in
scipy==1.5.4              # via scikit-learn, xgboost
six==1.12.0               # via absl-py, astunparse, cycler, ebooklib, google-auth, google-pasta, grpcio, h5py, imapclient, keras-preprocessing, langdetect, pdfminer.six, plotly, protobuf, python-dateutil, retrying, sacremoses, tensorboard, tensorflow, textract, traitlets, transliterate
sortedcontainers==2.3.0   # via pdfminer.six
soupsieve==2.0.1          # via beautifulsoup4
spacy==2.3.2              # via -r requirements.in
speechrecognition==3.8.1  # via textract
sqlparse==0.4.1           # via django
srsly==1.0.4              # via spacy, thinc
subword-nmt==0.3.7        # via laserembeddings
tablib[html,ods,xls,xlsx,yaml]==2.0.0  # via django-import-export
tensorboard-plugin-wit==1.7.0  # via tensorboard
tensorboard==2.4.0        # via -r requirements.in, tensorflow
tensorflow-estimator==2.3.0  # via tensorflow
tensorflow-hub==0.10.0    # via -r requirements.in
tensorflow==2.3.1         # via -r requirements.in
termcolor==1.1.0          # via tensorflow
textract==1.6.3           # via -r requirements.in
textsearch==0.0.17        # via contractions
textstat==0.6.2           # via -r requirements.in
thinc==7.4.1              # via spacy
threadpoolctl==2.1.0      # via scikit-learn
torch==1.7.0              # via laserembeddings
tqdm==4.53.0              # via nltk, sacremoses, spacy, thinc
traitlets==4.3.3          # via ipython
transliterate==1.10.2     # via laserembeddings
typing-extensions==3.7.4.3  # via torch
tzlocal==1.5.1            # via extract-msg
unidecode==1.1.1          # via textsearch
urllib3==1.26.2           # via requests
wasabi==0.8.0             # via spacy, thinc
wcwidth==0.2.5            # via prompt-toolkit
werkzeug==1.0.1           # via tensorboard
wheel==0.35.1             # via astunparse, tensorboard, tensorflow
whitenoise==5.2.0         # via -r requirements.in
wrapt==1.12.1             # via tensorflow
xgboost==1.2.1            # via -r requirements.in
xlrd==1.2.0               # via tablib, textract
xlsxwriter==1.3.7         # via python-pptx
xlwt==1.3.0               # via tablib
zipp==3.4.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
